refactor(boundaries): extract project_planning shared module (refs #1363)
The cockpit's Accept/Reject keybindings for improvement-proposal inbox
rows (cockpit_ui.action_accept_proposal / _finish_reject_proposal)
lazy-imported `accept_proposal`, `memkey_from_labels`, and
`record_proposal_rejection` from `pollypm.plugins_builtin.project_planning`.
That made the optional plugin effectively non-optional: disabling it
would silently break a core UI path.

Move those helpers (plus their `is_proposal_rejected` predicate pair
and the `REJECTIONS_FILE` path constant) into a new
`pollypm.project_planning_protocol` shared module. Both `proposals.py`
and `memory.py` re-export the names so existing imports (tests,
planner code, CLI) keep working unchanged. `cockpit_ui` now imports
straight from the protocol module — zero `plugins_builtin` imports
remain in that file.

No behaviour change — implementations are copied verbatim and identity
is preserved across the re-export boundary (verified by smoke test).
